% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmu}
\alias{cmu}
\title{CMU Pronouncing Dictionary}
\format{
A data frame with columns \code{word}, \code{pronunciation},
\code{phonemes_stressed}, and \code{phonemes}.
}
\source{
\url{http://www.speech.cs.cmu.edu/cgi-bin/cmudict}
}
\usage{
cmu
}
\description{
A data frame of words and their phonemic transcriptions from the Carnegie
Mellon University Pronouncing Dictionary, used for syllabification and
sonority scoring.
}
\keyword{datasets}
