% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmu_biphone_nonpos_cache}
\alias{cmu_biphone_nonpos_cache}
\title{CMU biphone frequency cache (non-positional, type)}
\format{
A list with \code{by_length}, \code{overall}, and \code{length_counts}.
}
\usage{
cmu_biphone_nonpos_cache
}
\description{
CMU biphone frequency cache (non-positional, type)
}
\keyword{internal}
