% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmu_triphone_cache}
\alias{cmu_triphone_cache}
\title{CMU triphone frequency cache (positional, type)}
\format{
A list with \code{matrices} and \code{length_counts}.
}
\usage{
cmu_triphone_cache
}
\description{
CMU triphone frequency cache (positional, type)
}
\keyword{internal}
