% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{count_moves}
\alias{count_moves}
\title{Find shortest number of moves to solution word.}
\usage{
count_moves(solution, anagram)
}
\arguments{
\item{solution}{The solution word string format.}

\item{anagram}{Anagram of the solution word in string format.}
}
\value{
Numeric length of minimum moves to solve.
}
\description{
Find shortest number of moves to solution word.
}
