% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ed1}
\alias{ed1}
\title{Compute the number of orthographic neighbours at edit distance 1.}
\usage{
ed1(word, dictionary = NULL, method = "lv")
}
\arguments{
\item{word}{A word or anagram in string format.}

\item{dictionary}{Character vector of dictionary words to compare against.}

\item{method}{Distance method passed to stringdist. Default is "lv".}
}
\value{
Numeric count of neighbours at edit distance 1.
}
\description{
Convenience wrapper around ed_n() with n = 1.
}
