% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{filter_anagrams}
\alias{filter_anagrams}
\title{Filter anagrams based on balance metrics}
\usage{
filter_anagrams(anagrams)
}
\arguments{
\item{anagrams}{Anagram pool from select_anagrams()}
}
\description{
Filter anagrams based on balance metrics
}
