% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{filter_unsolvable_anagrams}
\alias{filter_unsolvable_anagrams}
\title{Filter and select unsolvable anagrams}
\usage{
filter_unsolvable_anagrams(unsolvables)
}
\arguments{
\item{unsolvables}{Data frame from generate_unsolvable_anagrams()}
}
\description{
Filter and select unsolvable anagrams
}
