% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{first_letter}
\alias{first_letter}
\title{Identify whether the first letter of a word is a vowel or consonant.}
\usage{
first_letter(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Character string, either "vowel" or "consonant".
}
\description{
Identify whether the first letter of a word is a vowel or consonant.
}
