% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcide}
\alias{gcide}
\title{GCIDE lexicon}
\format{
A data frame with columns \code{word} and \code{pos}.
}
\usage{
gcide
}
\description{
A data frame of words and parts of speech from the GNU Collaborative
International Dictionary of English, used as one of four sources for
building the internal dictionary.
}
\keyword{datasets}
