% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{get_plottable_vars}
\alias{get_plottable_vars}
\title{Get plottable variables from a dataset}
\usage{
get_plottable_vars(data, type = c("solution", "anagram"))
}
\arguments{
\item{data}{Dataset with variables}

\item{type}{Either "solution" or "anagram"}
}
\description{
Get plottable variables from a dataset
}
