% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{grams_database_path}
\alias{grams_database_path}
\title{Get the local path to the GRAMS database}
\usage{
grams_database_path()
}
\value{
A character string giving the full path to \code{GRAMS_database.rds}.
}
\description{
Returns the path where the GRAMS database is stored in the user's app data
directory, creating the directory if it does not exist. This is the default
location used by \code{\link{download_grams_database}} and
\code{\link{load_grams_database}}.
}
\seealso{
\code{\link{download_grams_database}}, \code{\link{load_grams_database}}
}
