% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurtlex}
\alias{hurtlex}
\title{HurtLex offensive lexicon}
\format{
A data frame with columns \code{id}, \code{pos}, \code{category},
\code{stereotype}, \code{lemma}, and \code{level}.
}
\source{
\url{https://github.com/valeriobasile/hurtlex}
}
\usage{
hurtlex
}
\description{
A lexicon of harmful and offensive words used for censoring stimulus
candidates. From Bassignana et al. (2018).
}
\keyword{datasets}
