% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{infreq_letter}
\alias{infreq_letter}
\title{Identify whether the word contains an infrequent letter: "J", "K", "Q", "V",
"W", "X", or "Z".}
\usage{
infreq_letter(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Character string, either "vowel" or "consonant".
}
\description{
Identify whether the word contains an infrequent letter: "J", "K", "Q", "V",
"W", "X", or "Z".
}
