% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{intact_letters}
\alias{intact_letters}
\title{Finds letters that are the same in solution and anagram at the same position.}
\usage{
intact_letters(solution, anagram, output = "letters")
}
\arguments{
\item{solution}{The solution word in string format.}

\item{anagram}{Anagram of the solution word in string format.}

\item{output}{One of "letters" or "indices". Default "letters".}
}
\value{
Character vector of intact letters (uppercase) in left-to-right order
if output = "letters", or numeric vector of positions if output = "indices".
}
\description{
Finds letters that are the same in solution and anagram at the same position.
}
