% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ldnoobw}
\alias{ldnoobw}
\title{LDNOOBW offensive word list}
\format{
A data frame with column \code{word}.
}
\source{
\url{https://github.com/LDNOOBW/List-of-Dirty-Naughty-Obscene-and-Otherwise-Bad-Words}
}
\usage{
ldnoobw
}
\description{
A list of offensive words used for censoring stimulus candidates.
"List of Dirty, Naughty, Obscene, and Otherwise Bad Words."
}
\keyword{datasets}
