% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_grams_database}
\alias{load_grams_database}
\title{Load the GRAMS database}
\usage{
load_grams_database()
}
\value{
A data frame containing the GRAMS database.
}
\description{
Loads the GRAMS database from the user's app data directory. If the database
has not been downloaded yet, an informative error is raised with instructions.
}
\examples{
\dontrun{
GRAMS_database <- load_grams_database()
}

}
\seealso{
\code{\link{download_grams_database}}, \code{\link{grams_database_path}}
}
