% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morpheme_list}
\alias{morpheme_list}
\title{Morpheme list}
\format{
A tibble with 132 rows and 2 columns:
\describe{
\item{morpheme}{The morpheme string.}
\item{type}{Morpheme type, currently \code{"suffix"}.}
}
}
\usage{
morpheme_list
}
\description{
A tibble of common English morphemes and their types, used by
\code{find_morphemes} to identify morphological structure in words.
}
\keyword{datasets}
