% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{n_consonants}
\alias{n_consonants}
\title{Count the number of consonants in a word.}
\usage{
n_consonants(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Numeric count of consonants.
}
\description{
Count the number of consonants in a word.
}
