% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{n_unique_letters}
\alias{n_unique_letters}
\title{Count the number of unique letters in a word.}
\usage{
n_unique_letters(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Numeric count of unique letters.
}
\description{
Count the number of unique letters in a word.
}
