% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{n_vowels}
\alias{n_vowels}
\title{Count the number of vowels in a word.}
\usage{
n_vowels(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Numeric count of vowels.
}
\description{
Count the number of vowels in a word.
}
