% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthographic.R
\name{old_n}
\alias{old_n}
\title{Compute Orthographic Levenshtein Distance for N nearest neighbours.}
\usage{
old_n(word, n = 20, dictionary = subtlex_uk$word, method = "lv")
}
\arguments{
\item{word}{A word or anagram in string format.}

\item{n}{Number of nearest neighbours. Default is 20.}

\item{dictionary}{Character vector of dictionary words to compare against.}

\item{method}{Distance method passed to stringdist. Default is "lv".}
}
\value{
Numeric OLD-N value.
}
\description{
Compute Orthographic Levenshtein Distance for N nearest neighbours.
}
