% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonetics.R
\name{phonetics_setup}
\alias{phonetics_setup}
\title{Set up the phonetics Python environment.}
\usage{
phonetics_setup(envname = .phonetics_envname)
}
\arguments{
\item{envname}{Name of the virtual environment. Default is "r-phonetics".}
}
\value{
Invisible TRUE on success.
}
\description{
Creates a virtual environment and installs g2p_en, gruut, nltk, and torch,
then downloads required NLTK data. Only needs to be run once.
}
