% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{pld_n}
\alias{pld_n}
\title{Compute Phonological Levenshtein Distance for N nearest neighbours}
\usage{
pld_n(phonemes, n = 20, lexicon = NULL, method = "lv")
}
\arguments{
\item{phonemes}{Either a character vector of ARPAbet tokens or a
space-separated ARPAbet string.}

\item{n}{Number of nearest neighbours. Default is \code{20}.}

\item{lexicon}{Character vector of pre-encoded phoneme strings to compare
against. Defaults to \code{grams::cmu$phonemes_1char}.}

\item{method}{Distance method passed to \code{stringdist}. Default is
\code{"lv"}.}
}
\value{
Numeric PLD-N value.
}
\description{
Compute Phonological Levenshtein Distance for N nearest neighbours
}
