% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{plot_anagrams}
\alias{plot_anagrams}
\title{Plot anagram properties}
\usage{
plot_anagrams(
  anagrams,
  vars = c("aSBF", "aOLD20", "Moves"),
  show_labels = FALSE
)
}
\arguments{
\item{anagrams}{Anagram data with anagram_condition}

\item{vars}{Variables to plot (default: aSBF, aOLD20, Moves)}

\item{show_labels}{Show item labels on lines (default: FALSE)}
}
\description{
Plot anagram properties
}
