% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{plot_solutions}
\alias{plot_solutions}
\title{Plot solution word distributions by condition}
\usage{
plot_solutions(data, vars = NULL, show_means = FALSE, show_labels = FALSE)
}
\arguments{
\item{data}{LexOPS output (long format)}

\item{vars}{Variables to plot (default: all numeric except item_nr)}

\item{show_means}{Show mean values (default: FALSE)}

\item{show_labels}{Show item labels on lines (default: FALSE)}
}
\description{
Plot solution word distributions by condition
}
