% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{plot_unsolvables}
\alias{plot_unsolvables}
\title{Plot unsolvable anagram matching}
\usage{
plot_unsolvables(unsolvables, vars = c("SBF"), show_labels = FALSE)
}
\arguments{
\item{unsolvables}{Data frame with Anagram and Unsolvable columns plus metrics}

\item{vars}{Character vector of variables to plot}

\item{show_labels}{Logical, whether to show extreme value labels}
}
\description{
Plot unsolvable anagram matching
}
