% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{preserved_bg}
\alias{preserved_bg}
\title{Finds which bigrams appear in both solution and anagram.}
\usage{
preserved_bg(solution, anagram, output = "bigrams")
}
\arguments{
\item{solution}{The solution word in string format.}

\item{anagram}{Anagram of the solution word in string format.}

\item{output}{One of "bigrams" or "indices". Default "bigrams".}
}
\value{
Character vector of preserved bigrams (uppercase) in left-to-right
order if output = "bigrams", or numeric vector of starting positions if
output = "indices".
}
\description{
Finds which bigrams appear in both solution and anagram.
}
