% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{review_anagrams}
\alias{review_anagrams}
\title{Review and finalize anagram selection with plots}
\usage{
review_anagrams(anagrams)
}
\arguments{
\item{anagrams}{Filtered anagram pool from filter_anagrams()}
}
\description{
Review and finalize anagram selection with plots
}
