% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{review_unsolvable_anagrams}
\alias{review_unsolvable_anagrams}
\title{Review and finalize unsolvable anagram selection with plots}
\usage{
review_unsolvable_anagrams(unsolvables)
}
\arguments{
\item{unsolvables}{Filtered unsolvable anagrams from filter_unsolvable_anagrams()}
}
\description{
Review and finalize unsolvable anagram selection with plots
}
