% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtlex_uk_bigram_cache}
\alias{subtlex_uk_bigram_cache}
\title{SUBTLEX-UK bigram frequency cache (positional, type)}
\format{
A list with \code{matrices} and \code{length_counts}.
}
\usage{
subtlex_uk_bigram_cache
}
\description{
SUBTLEX-UK bigram frequency cache (positional, type)
}
\keyword{internal}
