% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtlex_uk_trigram_nonpos_cache}
\alias{subtlex_uk_trigram_nonpos_cache}
\title{SUBTLEX-UK trigram frequency cache (non-positional, type)}
\format{
A list with \code{by_length}, \code{overall}, and \code{length_counts}.
}
\usage{
subtlex_uk_trigram_nonpos_cache
}
\description{
SUBTLEX-UK trigram frequency cache (non-positional, type)
}
\keyword{internal}
