% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtlex_us}
\alias{subtlex_us}
\title{SUBTLEX-US frequency norms}
\format{
A data frame with columns including \code{word}, \code{Zipf},
\code{FREQcount}, \code{CDcount}, \code{Dom_PoS_SUBTLEX}, and others.
}
\source{
\url{http://crr.ugent.be/archives/4}
}
\usage{
subtlex_us
}
\description{
Word frequency norms from the SUBTLEX-US corpus (Brysbaert & New, 2009).
}
\keyword{datasets}
