% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{unique_letter_ratio}
\alias{unique_letter_ratio}
\title{Compute the ratio of unique letters to total letters in a word.}
\usage{
unique_letter_ratio(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Numeric ratio of distinct letters to word length.
}
\description{
Compute the ratio of unique letters to total letters in a word.
}
