% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{vowel_ratio}
\alias{vowel_ratio}
\title{Compute the ratio of vowels to total letters in a word.}
\usage{
vowel_ratio(word)
}
\arguments{
\item{word}{A word or anagram in string format.}
}
\value{
Numeric ratio of vowels to word length.
}
\description{
Compute the ratio of vowels to total letters in a word.
}
