% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordnet}
\alias{wordnet}
\title{WordNet lexicon}
\format{
A data frame with column \code{word}.
}
\source{
\url{https://wordnet.princeton.edu}
}
\usage{
wordnet
}
\description{
A character vector of words from the WordNet lexical database, used as
one of four sources for building the internal dictionary.
}
\keyword{datasets}
