% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods.R
\name{ed_n}
\alias{ed_n}
\title{Compute the number of orthographic neighbours at edit distance N.}
\usage{
ed_n(word, n = 1, dictionary = NULL, method = "lv")
}
\arguments{
\item{word}{A word or anagram in string format.}

\item{n}{Edit distance threshold. Default is 1.}

\item{dictionary}{Character vector of dictionary words to compare against.}

\item{method}{Distance method passed to stringdist. Default is "lv".}
}
\value{
Numeric count of neighbours at distance N.
}
\description{
Compute the number of orthographic neighbours at edit distance N.
}
