% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthographic.R
\name{orthographic_neighbours_list}
\alias{orthographic_neighbours_list}
\title{Return orthographic neighbours of a word at edit distance 1.}
\usage{
orthographic_neighbours_list(word, dictionary = subtlex_uk$word, method = "lv")
}
\arguments{
\item{word}{A word or anagram in string format.}

\item{dictionary}{Character vector of dictionary words to compare against.}

\item{method}{Distance method passed to stringdist. Default is "lv".}
}
\value{
Character vector of neighbouring words.
}
\description{
Return orthographic neighbours of a word at edit distance 1.
}
